Added error nesting correctly
Function now should get a nested error only in bind when the bound function returns an error, not when its input is an error. Errors in the input just get passed as is. Apply also got its error removed as the function that is passed into apply cannot create an error.